Beans Prediction
An agricultural scientist collected over 13,000 observations of physical properties of beans seeds (e.g. area, perimeter, axis lengths, etc.) from 7 different varieties of beans: Babunya, Bombay, Cali, Dermason, Horoz, Seker, and Sira. The data set contains the measurements for the collection of 13,000+ beans with the classification for the type of bean.
Using this data set, I partition the beans data set so that a random sample of 80% of the data will be used for training and 20% will be used for testing in the machine learning model. Using the test data I generated a multi-class confusion matrix to show the accuracy of the model.
I now load in a new test data set with random variables. Using the machine learning model I generated I can accurately predict the species of bean based on the data provided.